Refactored to organize better, and remove lots of get_region stuff from ER
