+++ worlds/tunic/ER_Data.py
@@ -1,0 +1,912 @@
+from typing import Dict, NamedTuple, List, Tuple
+
+
+class Portal(NamedTuple):
+    name: str  # human-readable name
+    region: str  # AP region
+    destination: str  # vanilla destination scene
+    tag: str  # vanilla destination tag, empty string if no tag
+    dead_end: bool = False  # dead end means there is only one exit from the room
+
+    # dead ends need to be prevented from connecting to each other
+
+    def scene(self) -> str:  # the actual scene name in Tunic
+        return tunic_er_regions[self.region].game_scene
+
+    def scene_destination_tag(self) -> str:  # full, nonchanging name to interpret by the mod
+        return self.scene() + ", " + self.destination + "_" + self.tag
+
+
+portal_mapping: List[Portal] = [
+    Portal(name="Stick House Entrance", region="Overworld",
+           destination="Sword Cave", tag=""),
+    Portal(name="Windmill Entrance", region="Overworld",
+           destination="Windmill", tag=""),
+    Portal(name="Well Ladder Entrance", region="Overworld",
+           destination="Sewer", tag="entrance"),
+    Portal(name="Entrance to Well from Well Rail", region="Overworld Well to Furnace Rail",
+           destination="Sewer", tag="west_aqueduct"),
+    Portal(name="Old House Entry Door", region="Overworld Old House Door",
+           destination="Overworld Interiors", tag="house"),
+    Portal(name="Old House Waterfall Entrance", region="Overworld",
+           destination="Overworld Interiors", tag="under_checkpoint"),
+    Portal(name="Entrance to Furnace from Well Rail", region="Overworld Well to Furnace Rail",
+           destination="Furnace", tag="gyro_upper_north"),
+    Portal(name="Entrance to Furnace from Windmill", region="Overworld",
+           destination="Furnace", tag="gyro_upper_east"),
+    Portal(name="Entrance to Furnace from West Garden", region="Overworld to West Garden from Furnace",
+           destination="Furnace", tag="gyro_west"),
+    Portal(name="Entrance to Furnace from Beach", region="Overworld",
+           destination="Furnace", tag="gyro_lower"),
+    Portal(name="Rotating Lights Entrance", region="Overworld",
+           destination="Overworld Cave", tag=""),
+    Portal(name="Swamp Upper Entrance", region="Overworld Laurels",
+           destination="Swamp Redux 2", tag="wall"),
+    Portal(name="Swamp Lower Entrance", region="Overworld",
+           destination="Swamp Redux 2", tag="conduit"),
+    Portal(name="Ruined Hall Entrance Not-Door", region="Overworld",
+           destination="Ruins Passage", tag="east"),
+    Portal(name="Ruined Hall Entrance Door", region="Overworld Ruined Hall Door",
+           destination="Ruins Passage", tag="west"),
+    Portal(name="Atoll Upper Entrance", region="Overworld",
+           destination="Atoll Redux", tag="upper"),
+    Portal(name="Atoll Lower Entrance", region="Overworld",
+           destination="Atoll Redux", tag="lower"),
+    Portal(name="Special Shop Entrance", region="Overworld Laurels",
+           destination="ShopSpecial", tag=""),
+    Portal(name="Maze Cave Entrance", region="Overworld",
+           destination="Maze Room", tag=""),
+    Portal(name="West Garden Entrance by Belltower", region="Overworld Belltower",
+           destination="Archipelagos Redux", tag="upper"),
+    Portal(name="West Garden Entrance by Dark Tomb", region="Overworld to West Garden from Furnace",
+           destination="Archipelagos Redux", tag="lower"),
+    Portal(name="West Garden Laurel Entrance", region="Overworld Laurels",
+           destination="Archipelagos Redux", tag="lowest"),
+    Portal(name="Temple Door Entrance", region="Overworld Temple Door",
+           destination="Temple", tag="main"),
+    Portal(name="Temple Rafters Entrance", region="Overworld",
+           destination="Temple", tag="rafters"),
+    Portal(name="Ruined Shop Entrance", region="Overworld",
+           destination="Ruined Shop", tag=""),
+    Portal(name="Patrol Cave Entrance", region="Overworld",
+           destination="PatrolCave", tag=""),
+    Portal(name="Hourglass Cave Entrance", region="Overworld",
+           destination="Town Basement", tag="beach"),
+    Portal(name="Changing Room Entrance", region="Overworld",
+           destination="Changing Room", tag=""),
+    Portal(name="Cube Room Entrance", region="Overworld",
+           destination="CubeRoom", tag=""),
+    Portal(name="Stairs from Overworld to Mountain", region="Overworld",
+           destination="Mountain", tag=""),
+    Portal(name="Overworld to Fortress", region="Overworld",
+           destination="Fortress Courtyard", tag=""),
+    Portal(name="Fountain HC Entrance", region="Overworld Fountain Cross Door",
+           destination="Town_FiligreeRoom", tag=""),
+    Portal(name="Glass Cannon HC Room Entrance", region="Overworld Southeast Cross Door",
+           destination="EastFiligreeCache", tag=""),
+    Portal(name="Overworld to Quarry Connector", region="Overworld",
+           destination="Darkwoods Tunnel", tag=""),
+    Portal(name="Dark Tomb Main Entrance", region="Overworld",
+           destination="Crypt Redux", tag=""),
+    Portal(name="Overworld to Forest Belltower", region="Overworld",
+           destination="Forest Belltower", tag=""),
+    Portal(name="Town Portal", region="Overworld Town Portal",
+           destination="Transit", tag="teleporter_town"),
+    Portal(name="Spawn Portal", region="Overworld Spawn Portal",
+           destination="Transit", tag="teleporter_starting island"),
+    Portal(name="Entrance to Fairy Cave", region="Overworld",
+           destination="Waterfall", tag=""),
+    Portal(name="Fairy Cave Exit", region="Secret Gathering Place",
+           destination="Overworld Redux", tag="",
+           dead_end=True),
+    Portal(name="Windmill Exit", region="Windmill",
+           destination="Overworld Redux", tag=""),
+    Portal(name="Windmill Shop", region="Windmill",
+           destination="Shop", tag=""),
+    Portal(name="Exit from Old House Front Door", region="Old House Front",
+           destination="Overworld Redux", tag="house"),
+    Portal(name="Teleport to Secret Treasure Room", region="Old House Front",
+           destination="g_elements", tag=""),
+    Portal(name="Exit from Old House Back Door", region="Old House Back",
+           destination="Overworld Redux", tag="under_checkpoint"),
+    Portal(name="Secret Treasure Room Exit", region="Relic Tower",
+           destination="Overworld Interiors", tag="",
+           dead_end=True),
+    Portal(name="Changing Room Exit", region="Changing Room",
+           destination="Overworld Redux", tag="",
+           dead_end=True),
+    Portal(name="Fountain HC Room Exit", region="Fountain Cross Room",
+           destination="Overworld Redux", tag="",
+           dead_end=True),
+    Portal(name="Cube Room Exit", region="Cube Cave",
+           destination="Overworld Redux", tag="",
+           dead_end=True),
+    Portal(name="Guard Patrol Cave Exit", region="Patrol Cave",
+           destination="Overworld Redux", tag="",
+           dead_end=True),
+    Portal(name="Ruined Shop Exit", region="Ruined Shop",
+           destination="Overworld Redux", tag="",
+           dead_end=True),
+    Portal(name="Furnace to Well Rail", region="Furnace Fuse",
+           destination="Overworld Redux", tag="gyro_upper_north"),
+    Portal(name="Furnace to Dark Tomb", region="Furnace Walking Path",
+           destination="Crypt Redux", tag=""),
+    Portal(name="Furnace to West Garden", region="Furnace Walking Path",
+           destination="Overworld Redux", tag="gyro_west"),
+    Portal(name="Furnace to Beach", region="Furnace Ladder Area",
+           destination="Overworld Redux", tag="gyro_lower"),
+    Portal(name="Furnace to Windmill", region="Furnace Ladder Area",
+           destination="Overworld Redux", tag="gyro_upper_east"),
+    Portal(name="Stick House Exit", region="Stick House",
+           destination="Overworld Redux", tag="",
+           dead_end=True),
+    Portal(name="Ruined Passage Door Exit", region="Ruined Passage",
+           destination="Overworld Redux", tag="east"),
+    Portal(name="Ruined Passage Not-door Exit", region="Ruined Passage",
+           destination="Overworld Redux", tag="west"),
+    Portal(name="Glass Cannon HC Room Exit", region="Southeast Cross Room",
+           destination="Overworld Redux", tag="",
+           dead_end=True),
+    Portal(name="Rotating Lights Exit", region="Caustic Light Cave",
+           destination="Overworld Redux", tag="",
+           dead_end=True),
+    Portal(name="Maze Cave Exit", region="Maze Cave",
+           destination="Overworld Redux", tag="",
+           dead_end=True),
+    Portal(name="Hourglass Cave Exit", region="Hourglass Cave",
+           destination="Overworld Redux", tag="beach",
+           dead_end=True),
+    Portal(name="Special Shop Exit", region="Special Shop",
+           destination="Overworld Redux", tag="",
+           dead_end=True),
+    Portal(name="Temple Rafters Exit", region="Sealed Temple Rafters",
+           destination="Overworld Redux", tag="rafters"),
+    Portal(name="Temple Door Exit", region="Sealed Temple",
+           destination="Overworld Redux", tag="main"),
+    Portal(name="Well Ladder Exit", region="Bottom of the Well Front",
+           destination="Overworld Redux", tag="entrance"),
+    Portal(name="Well to Well Boss", region="Bottom of the Well Back",
+           destination="Sewer_Boss", tag=""),
+    Portal(name="Well Rail Exit", region="Bottom of the Well Back",
+           destination="Overworld Redux", tag="west_aqueduct"),
+    Portal(name="Well Boss to Well", region="Well Boss",
+           destination="Sewer", tag=""),
+    Portal(name="Checkpoint to Dark Tomb", region="Dark Tomb Checkpoint",
+           destination="Crypt Redux", tag=""),
+    Portal(name="Dark Tomb to Overworld", region="Dark Tomb Entry Point",
+           destination="Overworld Redux", tag=""),
+    Portal(name="Dark Tomb to Furnace", region="Dark Tomb Dark Exit",
+           destination="Furnace", tag=""),
+    Portal(name="Dark Tomb to Checkpoint", region="Dark Tomb Entry Point",
+           destination="Sewer_Boss", tag=""),
+    Portal(name="West Garden Exit near Hero's Grave", region="West Garden",
+           destination="Overworld Redux", tag="lower"),
+    Portal(name="Magic Dagger House Entrance", region="West Garden",
+           destination="archipelagos_house", tag=""),
+    Portal(name="West Garden after Boss", region="West Garden after Boss",
+           destination="Overworld Redux", tag="upper"),
+    Portal(name="West Garden Shop", region="West Garden",
+           destination="Shop", tag=""),
+    Portal(name="West Garden Laurels Exit", region="West Garden Laurels Exit",
+           destination="Overworld Redux", tag="lowest"),
+    Portal(name="West Garden Hero's Grave", region="West Garden Hero's Grave",
+           destination="RelicVoid", tag="teleporter_relic plinth"),
+    Portal(name="West Garden Portal", region="West Garden Portal Area",
+           destination="Transit", tag="teleporter_archipelagos_teleporter",
+           dead_end=True),
+    Portal(name="Magic Dagger House Exit", region="Magic Dagger House",
+           destination="Archipelagos Redux", tag="",
+           dead_end=True),
+    Portal(name="Upper Atoll Exit", region="Ruined Atoll",
+           destination="Overworld Redux", tag="upper"),
+    Portal(name="Lower Atoll Exit", region="Ruined Atoll Lower Entry Area",
+           destination="Overworld Redux", tag="lower"),
+    Portal(name="Atoll Shop", region="Ruined Atoll",
+           destination="Shop", tag=""),
+    Portal(name="Atoll Portal", region="Ruined Atoll Portal",
+           destination="Transit", tag="teleporter_atoll"),
+    Portal(name="Atoll to Library", region="Ruined Atoll",
+           destination="Library Exterior", tag=""),
+    Portal(name="Frog Eye Entrance", region="Ruined Atoll",
+           destination="Frog Stairs", tag="eye"),
+    Portal(name="Frog Mouth Entrance", region="Ruined Atoll Frog Mouth",
+           destination="Frog Stairs", tag="mouth"),
+    Portal(name="Frog Eye Exit", region="Frog's Domain Entry",
+           destination="Atoll Redux", tag="eye"),
+    Portal(name="Frog Mouth Exit", region="Frog's Domain Entry",
+           destination="Atoll Redux", tag="mouth"),
+    Portal(name="Upper Frog to Lower Frog Entrance", region="Frog's Domain Entry",
+           destination="frog cave main", tag="Entrance"),
+    Portal(name="Upper Frog to Lower Frog Exit", region="Frog's Domain Entry",
+           destination="frog cave main", tag="Exit"),
+    Portal(name="Lower Frog Ladder Exit", region="Frog's Domain",
+           destination="Frog Stairs", tag="Entrance"),
+    Portal(name="Lower Frog Orb Exit", region="Frog's Domain Back",
+           destination="Frog Stairs", tag="Exit"),
+    Portal(name="Library to Atoll", region="Library Exterior Tree",
+           destination="Atoll Redux", tag=""),
+    Portal(name="Library Entry Ladder", region="Library Exterior Ladder",
+           destination="Library Hall", tag=""),
+    Portal(name="Library Bookshelf Exit", region="Library Hall",
+           destination="Library Exterior", tag=""),
+    Portal(name="Library Hero's Grave", region="Library Hero's Grave",
+           destination="RelicVoid", tag="teleporter_relic plinth"),
+    Portal(name="Lower Library to Rotunda", region="Library Hall",
+           destination="Library Rotunda", tag=""),
+    Portal(name="Library Rotunda Lower Exit", region="Library Rotunda",
+           destination="Library Hall", tag=""),
+    Portal(name="Library Rotunda Upper Exit", region="Library Rotunda",
+           destination="Library Lab", tag=""),
+    Portal(name="Upper Library to Rotunda", region="Library Lab Lower",
+           destination="Library Rotunda", tag=""),
+    Portal(name="Library Portal", region="Library Portal",
+           destination="Transit", tag="teleporter_library teleporter"),
+    Portal(name="Upper Library to Librarian", region="Library Lab",
+           destination="Library Arena", tag=""),
+    Portal(name="Librarian Arena Exit", region="Library Arena",
+           destination="Library Lab", tag="",
+           dead_end=True),
+    Portal(name="Forest to Belltower", region="East Forest",
+           destination="Forest Belltower", tag=""),
+    Portal(name="Forest Guard House 1 Lower Entrance", region="East Forest",
+           destination="East Forest Redux Laddercave", tag="lower"),
+    Portal(name="Forest Guard House 1 Gate Entrance", region="East Forest",
+           destination="East Forest Redux Laddercave", tag="gate"),
+    Portal(name="Forest Fox Dance Outside Doorway", region="East Forest Dance Fox Spot",
+           destination="East Forest Redux Laddercave", tag="upper"),
+    Portal(name="Forest Portal", region="East Forest Portal",
+           destination="Transit", tag="teleporter_forest teleporter"),
+    Portal(name="Forest Guard House 2 Lower Entrance", region="East Forest",
+           destination="East Forest Redux Interior", tag="lower"),
+    Portal(name="Forest Guard House 2 Upper Entrance", region="East Forest",
+           destination="East Forest Redux Interior", tag="upper"),
+    Portal(name="Forest Grave Path Lower Entrance", region="East Forest",
+           destination="Sword Access", tag="lower"),
+    Portal(name="Forest Grave Path Upper Entrance", region="East Forest",
+           destination="Sword Access", tag="upper"),
+    Portal(name="Guard House 1 Dance Exit", region="Guard House 1 West",
+           destination="East Forest Redux", tag="upper"),
+    Portal(name="Guard House 1 Lower Exit", region="Guard House 1 West",
+           destination="East Forest Redux", tag="lower"),
+    Portal(name="Guard House 1 Upper Forest Exit", region="Guard House 1 East",
+           destination="East Forest Redux", tag="gate"),
+    Portal(name="Guard House 1 to Guard Captain Room", region="Guard House 1 East",
+           destination="Forest Boss Room", tag=""),
+    Portal(name="Upper Forest Grave Path Exit", region="Forest Grave Path Upper",
+           destination="East Forest Redux", tag="upper"),
+    Portal(name="Lower Forest Grave Path Exit", region="Forest Grave Path Main",
+           destination="East Forest Redux", tag="lower"),
+    Portal(name="East Forest Hero's Grave", region="Forest Hero's Grave",
+           destination="RelicVoid", tag="teleporter_relic plinth"),
+    Portal(name="Guard House 2 Lower Exit", region="Guard House 2",
+           destination="East Forest Redux", tag="lower"),
+    Portal(name="Guard House 2 Upper Exit", region="Guard House 2",
+           destination="East Forest Redux", tag="upper"),
+    Portal(name="Guard Captain Room Non-Gate Exit", region="Forest Boss Room",
+           destination="East Forest Redux Laddercave", tag=""),
+    Portal(name="Guard Captain Room Gate Exit", region="Forest Boss Room",
+           destination="Forest Belltower", tag=""),
+    Portal(name="Forest Belltower to Fortress", region="Forest Belltower Main",
+           destination="Fortress Courtyard", tag=""),
+    Portal(name="Forest Belltower to Forest", region="Forest Belltower Lower",
+           destination="East Forest Redux", tag=""),
+    Portal(name="Forest Belltower to Overworld", region="Forest Belltower Main",
+           destination="Overworld Redux", tag=""),
+    Portal(name="Forest Belltower to Guard Captain Room", region="Forest Belltower Upper",
+           destination="Forest Boss Room", tag=""),
+    Portal(name="Lower Fortress Grave Path Entrance", region="Fortress Courtyard",
+           destination="Fortress Reliquary", tag="Lower"),
+    Portal(name="Upper Fortress Grave Path Entrance", region="Fortress Courtyard Upper",
+           destination="Fortress Reliquary", tag="Upper"),
+    Portal(name="Fortress Courtyard to Fortress Interior", region="Fortress Courtyard",
+           destination="Fortress Main", tag="Big Door"),
+    Portal(name="Fortress Courtyard to Fortress East", region="Fortress Courtyard Upper",
+           destination="Fortress East", tag=""),
+    Portal(name="Fortress Courtyard to Beneath the Earth", region="Fortress Exterior near cave",
+           destination="Fortress Basement", tag=""),
+    Portal(name="Fortress Courtyard to Forest Belltower", region="Fortress Exterior from East Forest",
+           destination="Forest Belltower", tag=""),
+    Portal(name="Fortress Courtyard to Overworld", region="Fortress Exterior from Overworld",
+           destination="Overworld Redux", tag=""),
+    Portal(name="Fortress Courtyard Shop", region="Fortress Exterior near cave",
+           destination="Shop", tag=""),
+    Portal(name="Beneath the Earth to Fortress Interior", region="Beneath the Vault Back",
+           destination="Fortress Main", tag=""),
+    Portal(name="Beneath the Earth to Fortress Courtyard", region="Beneath the Vault Front",
+           destination="Fortress Courtyard", tag=""),
+    Portal(name="Fortress Main Exit", region="Eastern Vault Fortress",
+           destination="Fortress Courtyard", tag="Big Door"),
+    Portal(name="Fortress Interior to Beneath the Earth", region="Eastern Vault Fortress",
+           destination="Fortress Basement", tag=""),
+    Portal(name="Fortress Interior to Siege Engine", region="Eastern Vault Fortress Gold Door",
+           destination="Fortress Arena", tag=""),
+    Portal(name="Fortress Interior Shop", region="Eastern Vault Fortress",
+           destination="Shop", tag=""),
+    Portal(name="Fortress Interior to East Fortress Upper", region="Eastern Vault Fortress",
+           destination="Fortress East", tag="upper"),
+    Portal(name="Fortress Interior to East Fortress Lower", region="Eastern Vault Fortress",
+           destination="Fortress East", tag="lower"),
+    Portal(name="East Fortress to Interior Lower", region="Fortress East Shortcut Lower",
+           destination="Fortress Main", tag="lower"),
+    Portal(name="East Fortress to Courtyard", region="Fortress East Shortcut Upper",
+           destination="Fortress Courtyard", tag=""),
+    Portal(name="East Fortress to Interior Upper", region="Fortress East Shortcut Upper",
+           destination="Fortress Main", tag="upper"),
+    Portal(name="Lower Fortress Grave Path Exit", region="Fortress Grave Path",
+           destination="Fortress Courtyard", tag="Lower"),
+    Portal(name="Fortress Hero's Grave", region="Fortress Grave Path",
+           destination="RelicVoid", tag="teleporter_relic plinth"),
+    Portal(name="Fortress Grave Path Dusty Entrance", region="Fortress Grave Path Dusty Entrance",
+           destination="Dusty", tag=""),
+    Portal(name="Dusty Exit", region="Fortress Leaf Piles",
+           destination="Fortress Reliquary", tag="",
+           dead_end=True),
+    Portal(name="Upper Fortress Grave Path Exit", region="Fortress Grave Path Upper",
+           destination="Fortress Courtyard", tag="Upper",
+           dead_end=True),
+    Portal(name="Siege Engine Arena to Fortress", region="Fortress Arena",
+           destination="Fortress Main", tag=""),
+    Portal(name="Fortress Portal", region="Fortress Arena Portal",
+           destination="Transit", tag="teleporter_spidertank"),
+    Portal(name="Stairs to Top of the Mountain", region="Lower Mountain Stairs",
+           destination="Mountaintop", tag=""),
+    Portal(name="Mountain to Quarry", region="Lower Mountain",
+           destination="Quarry Redux", tag=""),
+    Portal(name="Mountain to Overworld", region="Lower Mountain",
+           destination="Overworld Redux", tag=""),
+    Portal(name="Top of the Mountain Exit", region="Top of the Mountain",
+           destination="Mountain", tag="",
+           dead_end=True),
+    Portal(name="Quarry Connector to Overworld", region="Quarry Connector",
+           destination="Overworld Redux", tag=""),
+    Portal(name="Quarry Connector to Quarry", region="Quarry Connector",
+           destination="Quarry Redux", tag=""),
+    Portal(name="Quarry to Overworld Exit", region="Quarry",
+           destination="Darkwoods Tunnel", tag=""),
+    Portal(name="Quarry Shop", region="Quarry",
+           destination="Shop", tag=""),
+    Portal(name="Quarry to Monastery Front", region="Quarry",
+           destination="Monastery", tag="front"),
+    Portal(name="Quarry to Monastery Back", region="Monastery Rope",
+           destination="Monastery", tag="back"),
+    Portal(name="Quarry to Mountain", region="Quarry",
+           destination="Mountain", tag=""),
+    Portal(name="Quarry Zig Entrance", region="Lower Quarry Zig Door",
+           destination="ziggurat2020_0", tag=""),
+    Portal(name="Quarry Portal", region="Quarry Portal",
+           destination="Transit", tag="teleporter_quarry teleporter"),
+    Portal(name="Monastery Rear Exit", region="Monastery Back",
+           destination="Quarry Redux", tag="back"),
+    Portal(name="Monastery Front Exit", region="Monastery Front",
+           destination="Quarry Redux", tag="front"),
+    Portal(name="Monastery Hero's Grave", region="Monastery Hero's Grave",
+           destination="RelicVoid", tag="teleporter_relic plinth"),
+    Portal(name="Zig Entry Hallway to Zig 1", region="Rooted Ziggurat Entry",
+           destination="ziggurat2020_1", tag=""),
+    Portal(name="Zig Entry Hallway to Quarry", region="Rooted Ziggurat Entry",
+           destination="Quarry Redux", tag=""),
+    Portal(name="Zig 1 to Zig Entry", region="Rooted Ziggurat Upper Front",
+           destination="ziggurat2020_0", tag=""),
+    Portal(name="Zig 1 to Zig 2", region="Rooted Ziggurat Upper Back",
+           destination="ziggurat2020_2", tag=""),
+    Portal(name="Zig 2 to Zig 1", region="Rooted Ziggurat Middle Top",
+           destination="ziggurat2020_1", tag=""),
+    Portal(name="Zig 2 to Zig 3", region="Rooted Ziggurat Middle Bottom",
+           destination="ziggurat2020_3", tag=""),
+    Portal(name="Zig Portal Room Entrance", region="Rooted Ziggurat Portal Room Entrance",
+           destination="ziggurat2020_FTRoom", tag=""),
+    Portal(name="Zig 3 to Zig 2", region="Rooted Ziggurat Lower Front",
+           destination="ziggurat2020_2", tag=""),
+    Portal(name="Zig Portal Room Exit", region="Rooted Ziggurat Portal Room Exit",
+           destination="ziggurat2020_3", tag=""),
+    Portal(name="Zig Portal", region="Rooted Ziggurat Portal",
+           destination="Transit", tag="teleporter_ziggurat teleporter"),
+    Portal(name="Lower Swamp Exit", region="Swamp",
+           destination="Overworld Redux", tag="conduit"),
+    Portal(name="Swamp to Cathedral Main Entrance", region="Swamp to Cathedral Main Entrance",
+           destination="Cathedral Redux", tag="main"),
+    Portal(name="Swamp to Cathedral Treasure Room Entrance", region="Swamp to Cathedral Treasure Room",
+           destination="Cathedral Redux", tag="secret"),
+    Portal(name="Swamp to Gauntlet", region="Back of Swamp",
+           destination="Cathedral Arena", tag=""),
+    Portal(name="Swamp Shop", region="Swamp",
+           destination="Shop", tag=""),
+    Portal(name="Upper Swamp Exit", region="Back of Swamp Laurels Area",
+           destination="Overworld Redux", tag="wall"),
+    Portal(name="Swamp Hero's Grave", region="Swamp Hero's Grave",
+           destination="RelicVoid", tag="teleporter_relic plinth"),
+    Portal(name="Cathedral Main Exit", region="Cathedral",
+           destination="Swamp Redux 2", tag="main"),
+    Portal(name="Cathedral Elevator", region="Cathedral",
+           destination="Cathedral Arena", tag=""),
+    Portal(name="Cathedral Treasure Room Exit", region="Cathedral Secret Legend Room",
+           destination="Swamp Redux 2", tag="secret",
+           dead_end=True),
+    Portal(name="Gauntlet to Swamp", region="Cathedral Gauntlet Exit",
+           destination="Swamp Redux 2", tag="",
+           dead_end=True),
+    Portal(name="Gauntlet Elevator", region="Cathedral Gauntlet Checkpoint",
+           destination="Cathedral Redux", tag=""),
+    Portal(name="Gauntlet Shop", region="Cathedral Gauntlet Checkpoint",
+           destination="Shop", tag=""),
+    Portal(name="Hero Relic to Fortress", region="Hero Relic - Fortress",
+           destination="Fortress Reliquary", tag="teleporter_relic plinth",
+           dead_end=True),
+    Portal(name="Hero Relic to Monastery", region="Hero Relic - Quarry",
+           destination="Monastery", tag="teleporter_relic plinth",
+           dead_end=True),
+    Portal(name="Hero Relic to West Garden", region="Hero Relic - West Garden",
+           destination="Archipelagos Redux", tag="teleporter_relic plinth",
+           dead_end=True),
+    Portal(name="Hero Relic to East Forest", region="Hero Relic - East Forest",
+           destination="Sword Access", tag="teleporter_relic plinth",
+           dead_end=True),
+    Portal(name="Hero Relic to Library", region="Hero Relic - Library",
+           destination="Library Hall", tag="teleporter_relic plinth",
+           dead_end=True),
+    Portal(name="Hero Relic to Swamp", region="Hero Relic - Swamp",
+           destination="Swamp Redux 2", tag="teleporter_relic plinth",
+           dead_end=True),
+    Portal(name="Far Shore to West Garden", region="Far Shore to West Garden",
+           destination="Archipelagos Redux", tag="teleporter_archipelagos_teleporter"),
+    Portal(name="Far Shore to Library", region="Far Shore to Library",
+           destination="Library Lab", tag="teleporter_library teleporter"),
+    Portal(name="Far Shore to Quarry", region="Far Shore to Quarry",
+           destination="Quarry Redux", tag="teleporter_quarry teleporter"),
+    Portal(name="Far Shore to East Forest", region="Far Shore to East Forest",
+           destination="East Forest Redux", tag="teleporter_forest teleporter"),
+    Portal(name="Far Shore to Fortress", region="Far Shore to Fortress",
+           destination="Fortress Arena", tag="teleporter_spidertank"),
+    Portal(name="Far Shore to Atoll", region="Far Shore",
+           destination="Atoll Redux", tag="teleporter_atoll"),
+    Portal(name="Far Shore to Zig", region="Far Shore",
+           destination="ziggurat2020_FTRoom", tag="teleporter_ziggurat teleporter"),
+    Portal(name="Far Shore to Heir", region="Far Shore",
+           destination="Spirit Arena", tag="teleporter_spirit arena"),
+    Portal(name="Far Shore to Town", region="Far Shore",
+           destination="Overworld Redux", tag="teleporter_town"),
+    Portal(name="Far Shore to Spawn", region="Far Shore to Spawn",
+           destination="Overworld Redux", tag="teleporter_starting island"),
+    Portal(name="Heir Arena Exit", region="Spirit Arena",
+           destination="Transit", tag="teleporter_spirit arena",
+           dead_end=True),
+    Portal(name="Purgatory Bottom Exit", region="Purgatory",
+           destination="Purgatory", tag="bottom"),
+    Portal(name="Purgatory Top Exit", region="Purgatory",
+           destination="Purgatory", tag="top"),
+]
+
+
+class RegionInfo(NamedTuple):
+    game_scene: str  # the name of the scene in the actual game
+    isolated: bool = False  # if this region has 1 entrance, coincides with the portal dead_ends
+
+
+# first string is the AP Region name, second string is the Scene name in-game
+tunic_er_regions: Dict[str, RegionInfo] = {
+    "Menu": RegionInfo("Fake", True),
+    "Overworld": RegionInfo("Overworld Redux"),
+    "Overworld Holy Cross": RegionInfo("Fake", True),
+    "Overworld Belltower": RegionInfo("Overworld Redux"),  # the area with the belltower and chest
+    "Overworld Laurels": RegionInfo("Overworld Redux"),  # all spots in Overworld that you need laurels to reach
+    "Overworld to West Garden from Furnace": RegionInfo("Overworld Redux"),  # that little connector after Dark Tomb
+    "Overworld Well to Furnace Rail": RegionInfo("Overworld Redux"),  # the tiny rail passageway
+    "Overworld Ruined Hall Door": RegionInfo("Overworld Redux"),  # the small space betweeen the door and the portal
+    "Overworld Old House Door": RegionInfo("Overworld Redux"),  # the too-small space between the door and the portal
+    "Overworld Southeast Cross Door": RegionInfo("Overworld Redux"),  # the small space betweeen the door and the portal
+    "Overworld Fountain Cross Door": RegionInfo("Overworld Redux"),
+    "Overworld Temple Door": RegionInfo("Overworld Redux"),  # the small space betweeen the door and the portal
+    "Overworld Town Portal": RegionInfo("Overworld Redux"),
+    "Overworld Spawn Portal": RegionInfo("Overworld Redux"),
+    "Stick House": RegionInfo("Sword Cave", True),
+    "Windmill": RegionInfo("Windmill"),
+    "Old House Back": RegionInfo("Overworld Interiors"),  # part with the hc door
+    "Old House Front": RegionInfo("Overworld Interiors"),  # part with the bedroom
+    "Relic Tower": RegionInfo("g_elements", True),
+    "Furnace Fuse": RegionInfo("Furnace"),  # top of the furnace
+    "Furnace Ladder Area": RegionInfo("Furnace"),  # the two portals accessible by the ladder
+    "Furnace Walking Path": RegionInfo("Furnace"),  # dark tomb to west garden
+    "Secret Gathering Place": RegionInfo("Waterfall", True),
+    "Changing Room": RegionInfo("Changing Room", True),
+    "Patrol Cave": RegionInfo("PatrolCave", True),
+    "Ruined Shop": RegionInfo("Ruined Shop", True),
+    "Ruined Passage": RegionInfo("Ruins Passage"),
+    "Special Shop": RegionInfo("ShopSpecial", True),
+    "Caustic Light Cave": RegionInfo("Overworld Cave", True),
+    "Maze Cave": RegionInfo("Maze Room", True),
+    "Cube Cave": RegionInfo("CubeRoom", True),
+    "Southeast Cross Room": RegionInfo("EastFiligreeCache", True),
+    "Fountain Cross Room": RegionInfo("Town_FiligreeRoom", True),
+    "Hourglass Cave": RegionInfo("Town Basement", True),
+    "Sealed Temple": RegionInfo("Temple"),
+    "Sealed Temple Rafters": RegionInfo("Temple"),
+    "Forest Belltower Upper": RegionInfo("Forest Belltower"),
+    "Forest Belltower Main": RegionInfo("Forest Belltower"),
+    "Forest Belltower Lower": RegionInfo("Forest Belltower"),
+    "East Forest": RegionInfo("East Forest Redux"),
+    "East Forest Dance Fox Spot": RegionInfo("East Forest Redux"),
+    "East Forest Portal": RegionInfo("East Forest Redux"),
+    "Guard House 1 East": RegionInfo("East Forest Redux Laddercave"),
+    "Guard House 1 West": RegionInfo("East Forest Redux Laddercave"),
+    "Guard House 2": RegionInfo("East Forest Redux Interior"),
+    "Forest Boss Room": RegionInfo("Forest Boss Room"),
+    "Forest Grave Path Main": RegionInfo("Sword Access"),
+    "Forest Grave Path Upper": RegionInfo("Sword Access"),
+    "Forest Grave Path by Grave": RegionInfo("Sword Access"),
+    "Forest Hero's Grave": RegionInfo("Sword Access"),
+    "Dark Tomb Entry Point": RegionInfo("Crypt Redux"),  # top exit & checkpoint exit
+    "Dark Tomb Main": RegionInfo("Crypt Redux"),
+    "Dark Tomb Dark Exit": RegionInfo("Crypt Redux"),
+    "Dark Tomb Checkpoint": RegionInfo("Sewer_Boss"),  # can laurels backwards
+    "Well Boss": RegionInfo("Sewer_Boss"),  # can walk through (with bombs at least)
+    "Bottom of the Well Front": RegionInfo("Sewer"),
+    "Bottom of the Well Back": RegionInfo("Sewer"),  # add this in the other spots
+    "West Garden": RegionInfo("Archipelagos Redux"),
+    "Magic Dagger House": RegionInfo("archipelagos_house", True),
+    "West Garden Portal Area": RegionInfo("Archipelagos Redux", True),
+    "West Garden Laurels Exit": RegionInfo("Archipelagos Redux"),
+    "West Garden after Boss": RegionInfo("Archipelagos Redux"),
+    "West Garden Hero's Grave": RegionInfo("Archipelagos Redux"),
+    "Ruined Atoll": RegionInfo("Atoll Redux"),
+    "Ruined Atoll Lower Entry Area": RegionInfo("Atoll Redux"),
+    "Ruined Atoll Frog Mouth": RegionInfo("Atoll Redux"),
+    "Ruined Atoll Portal": RegionInfo("Atoll Redux"),
+    "Frog's Domain Entry": RegionInfo("Frog Stairs"),
+    "Frog's Domain": RegionInfo("frog cave main"),
+    "Frog's Domain Back": RegionInfo("frog cave main"),
+    "Library Exterior Tree": RegionInfo("Library Exterior"),
+    "Library Exterior Ladder": RegionInfo("Library Exterior"),
+    "Library Hall": RegionInfo("Library Hall"),
+    "Library Hero's Grave": RegionInfo("Library Hall"),
+    "Library Rotunda": RegionInfo("Library Rotunda"),
+    "Library Lab": RegionInfo("Library Lab"),
+    "Library Lab Lower": RegionInfo("Library Lab"),
+    "Library Portal": RegionInfo("Library Lab"),
+    "Library Arena": RegionInfo("Library Arena", True),
+    "Fortress Exterior from East Forest": RegionInfo("Fortress Courtyard"),
+    "Fortress Exterior from Overworld": RegionInfo("Fortress Courtyard"),
+    "Fortress Exterior near cave": RegionInfo("Fortress Courtyard"),  # where the shop and beneath the earth entry are
+    "Fortress Courtyard": RegionInfo("Fortress Courtyard"),
+    "Fortress Courtyard Upper": RegionInfo("Fortress Courtyard"),
+    "Beneath the Vault Front": RegionInfo("Fortress Basement"),  # the vanilla entry point
+    "Beneath the Vault Back": RegionInfo("Fortress Basement"),  # the vanilla exit point
+    "Eastern Vault Fortress": RegionInfo("Fortress Main"),
+    "Eastern Vault Fortress Gold Door": RegionInfo("Fortress Main"),
+    "Fortress East Shortcut Upper": RegionInfo("Fortress East"),
+    "Fortress East Shortcut Lower": RegionInfo("Fortress East"),
+    "Fortress Grave Path": RegionInfo("Fortress Reliquary"),
+    "Fortress Grave Path Upper": RegionInfo("Fortress Reliquary", True),
+    "Fortress Grave Path Dusty Entrance": RegionInfo("Fortress Reliquary"),
+    "Fortress Hero's Grave": RegionInfo("Fortress Reliquary"),
+    "Fortress Leaf Piles": RegionInfo("Dusty", True),
+    "Fortress Arena": RegionInfo("Fortress Arena"),
+    "Fortress Arena Portal": RegionInfo("Fortress Arena"),
+    "Lower Mountain": RegionInfo("Mountain"),
+    "Lower Mountain Stairs": RegionInfo("Mountain"),
+    "Top of the Mountain": RegionInfo("Mountaintop", True),
+    "Quarry Connector": RegionInfo("Darkwoods Tunnel"),
+    "Quarry": RegionInfo("Quarry Redux"),
+    "Quarry Portal": RegionInfo("Quarry Redux"),
+    "Monastery Front": RegionInfo("Monastery"),
+    "Monastery Back": RegionInfo("Monastery"),
+    "Monastery Hero's Grave": RegionInfo("Monastery"),
+    "Monastery Rope": RegionInfo("Quarry Redux"),
+    "Lower Quarry": RegionInfo("Quarry Redux"),
+    "Lower Quarry Zig Door": RegionInfo("Quarry Redux"),
+    "Rooted Ziggurat Entry": RegionInfo("ziggurat2020_0"),
+    "Rooted Ziggurat Upper Front": RegionInfo("ziggurat2020_1"),
+    "Rooted Ziggurat Upper Back": RegionInfo("ziggurat2020_1"),  # after the administrator
+    "Rooted Ziggurat Middle Top": RegionInfo("ziggurat2020_2"),
+    "Rooted Ziggurat Middle Bottom": RegionInfo("ziggurat2020_2"),
+    "Rooted Ziggurat Lower Front": RegionInfo("ziggurat2020_3"),  # the vanilla entry point side
+    "Rooted Ziggurat Lower Back": RegionInfo("ziggurat2020_3"),  # the boss side
+    "Rooted Ziggurat Portal Room Entrance": RegionInfo("ziggurat2020_3"),  # the door itself on the zig 3 side
+    "Rooted Ziggurat Portal": RegionInfo("ziggurat2020_FTRoom"),
+    "Rooted Ziggurat Portal Room Exit": RegionInfo("ziggurat2020_FTRoom"),
+    "Swamp": RegionInfo("Swamp Redux 2"),
+    "Swamp to Cathedral Treasure Room": RegionInfo("Swamp Redux 2"),
+    "Swamp to Cathedral Main Entrance": RegionInfo("Swamp Redux 2"),
+    "Back of Swamp": RegionInfo("Swamp Redux 2"),  # the area with hero grave and gauntlet entrance
+    "Swamp Hero's Grave": RegionInfo("Swamp Redux 2"),
+    "Back of Swamp Laurels Area": RegionInfo("Swamp Redux 2"),  # the spots you need laurels to traverse
+    "Cathedral": RegionInfo("Cathedral Redux"),
+    "Cathedral Secret Legend Room": RegionInfo("Cathedral Redux", True),
+    "Cathedral Gauntlet Checkpoint": RegionInfo("Cathedral Arena"),
+    "Cathedral Gauntlet": RegionInfo("Cathedral Arena"),
+    "Cathedral Gauntlet Exit": RegionInfo("Cathedral Arena"),
+    "Far Shore": RegionInfo("Transit"),
+    "Far Shore to Spawn": RegionInfo("Transit"),
+    "Far Shore to East Forest": RegionInfo("Transit"),
+    "Far Shore to Quarry": RegionInfo("Transit"),
+    "Far Shore to Fortress": RegionInfo("Transit"),
+    "Far Shore to Library": RegionInfo("Transit"),
+    "Far Shore to West Garden": RegionInfo("Transit"),
+    "Hero Relic - Fortress": RegionInfo("RelicVoid", True),
+    "Hero Relic - Quarry": RegionInfo("RelicVoid", True),
+    "Hero Relic - West Garden": RegionInfo("RelicVoid", True),
+    "Hero Relic - East Forest": RegionInfo("RelicVoid", True),
+    "Hero Relic - Library": RegionInfo("RelicVoid", True),
+    "Hero Relic - Swamp": RegionInfo("RelicVoid", True),
+    "Purgatory": RegionInfo("Purgatory"),
+    "Shop": RegionInfo("Shop", True),
+    "Spirit Arena": RegionInfo("Spirit Arena", True),
+    "Spirit Arena Victory": RegionInfo("Spirit Arena", True)
+}
+
+
+class StaticCxn(NamedTuple):
+    origin: str
+    destination: str
+    reqs: List[List[str]] = []
+    region_reqs: List[str] = []
+    reverse: bool = False  # if the reverse connection has the same requirements
+
+
+er_static_cxns: List[StaticCxn] = [
+    StaticCxn(origin="Menu", destination="Overworld"),
+    StaticCxn(origin="Overworld Belltower", destination="Overworld"),
+    StaticCxn(origin="Overworld", destination="Overworld Belltower",
+              reqs=[["Hero's Laurels"]]),
+    StaticCxn(origin="Overworld", destination="Overworld Holy Cross",
+              reqs=[["Holy Cross"]]),
+    StaticCxn(origin="Overworld", destination="Overworld Ruined Hall Door",
+              reqs=[["Key"]]),  # one-way because the reverse isn't necessary
+    StaticCxn(origin="Overworld Laurels", destination="Overworld",
+              reqs=[["Hero's Laurels"]], reverse=True),
+    StaticCxn(origin="Overworld", destination="Overworld Old House Door",
+              reqs=[["Old House Key"]]),
+    StaticCxn(origin="Overworld Southeast Cross Door", destination="Overworld",
+              reqs=[["Holy Cross"]], reverse=True),
+    StaticCxn(origin="Overworld Fountain Cross Door", destination="Overworld"),  # can exit with nothing
+    StaticCxn(origin="Overworld", destination="Overworld Fountain Cross Door",
+              reqs=[["Holy Cross"]]),  # needs holy cross to enter it
+    StaticCxn(origin="Overworld Town Portal", destination="Overworld"),  # can exit from it with nothing
+    StaticCxn(origin="Overworld", destination="Overworld Town Portal",
+              reqs=[["Prayer"]]),  # but needs prayer to enter it
+    StaticCxn(origin="Overworld Spawn Portal", destination="Overworld"),
+    StaticCxn(origin="Overworld", destination="Overworld Spawn Portal",
+              reqs=[["Prayer"]]),
+    StaticCxn(origin="Overworld Temple Door", destination="Overworld",
+              reqs=[["Stick"], ["Magic Wand"]],
+              region_reqs=["Overworld Belltower", "Forest Belltower Upper"], reverse=True),
+    StaticCxn(origin="Old House Front", destination="Old House Back"),
+    StaticCxn(origin="Sealed Temple", destination="Sealed Temple Rafters"),
+    StaticCxn(origin="Sealed Temple Rafters", destination="Sealed Temple",
+              reqs=[["Hero's Laurels"]]),
+    StaticCxn(origin="Furnace Ladder Area", destination="Furnace Walking Path",
+              reqs=[["Hero's Laurels"]], reverse=True),
+    StaticCxn(origin="Furnace Ladder Area", destination="Furnace Fuse",
+              reqs=[["Hero's Laurels"]], reverse=True),
+    StaticCxn(origin="Furnace Fuse", destination="Furnace Walking Path",
+              reqs=[["Hero's Laurels"]], reverse=True),
+    StaticCxn(origin="Forest Belltower Upper", destination="Forest Belltower Main"),
+    StaticCxn(origin="Forest Belltower Main", destination="Forest Belltower Lower"),
+    StaticCxn(origin="East Forest", destination="East Forest Dance Fox Spot",
+              reqs=[["Hero's Laurels"]], reverse=True),
+    StaticCxn(origin="East Forest", destination="East Forest Portal",
+              reqs=[["Prayer"]]),
+    StaticCxn(origin="East Forest Portal", destination="East Forest"),
+    StaticCxn(origin="Guard House 1 East", destination="Guard House 1 West"),
+    StaticCxn(origin="Guard House 1 West", destination="Guard House 1 East",
+              reqs=[["Hero's Laurels"]]),
+    StaticCxn(origin="Forest Grave Path Upper", destination="Forest Grave Path Main",
+              reqs=[["Hero's Laurels"]], reverse=True),
+    StaticCxn(origin="Forest Grave Path Main", destination="Forest Grave Path by Grave"),
+    StaticCxn(origin="Forest Grave Path by Grave", destination="Forest Hero's Grave",
+              reqs=[["Prayer"]]),
+    StaticCxn(origin="Forest Hero's Grave", destination="Forest Grave Path by Grave"),
+    StaticCxn(origin="Bottom of the Well Front", destination="Bottom of the Well Back",
+              reqs=[["Stick"]], reverse=True),
+    StaticCxn(origin="Well Boss", destination="Dark Tomb Checkpoint"),
+    StaticCxn(origin="Dark Tomb Checkpoint", destination="Well Boss",
+              reqs=[["Hero's Laurels"]]),
+    StaticCxn(origin="Dark Tomb Entry Point", destination="Dark Tomb Main",
+              reqs=[["Lantern"]], reverse=True),
+    StaticCxn(origin="Dark Tomb Main", destination="Dark Tomb Dark Exit",
+              reqs=[["Lantern"]], reverse=True),
+    StaticCxn(origin="West Garden Laurels Exit", destination="West Garden",
+              reqs=[["Hero's Laurels"]], reverse=True),
+    StaticCxn(origin="West Garden after Boss", destination="West Garden",
+              reqs=[["Hero's Laurels"]]),
+    StaticCxn(origin="West Garden", destination="West Garden after Boss",
+              reqs=[["Hero's Laurels"], ["Sword"]]),
+    StaticCxn(origin="West Garden", destination="West Garden Hero's Grave",
+              reqs=[["Prayer"]]),
+    StaticCxn(origin="West Garden Hero's Grave", destination="West Garden"),
+    StaticCxn(origin="Ruined Atoll", destination="Ruined Atoll Lower Entry Area",
+              reqs=[["Hero's Laurels"]]),
+    StaticCxn(origin="Ruined Atoll Frog Mouth", destination="Ruined Atoll",
+              reqs=[["Hero's Laurels"], ["Magic Orb"]], reverse=True),
+    StaticCxn(origin="Ruined Atoll Lower Entry Area", destination="Ruined Atoll",
+              reqs=[["Hero's Laurels"], ["Magic Orb"]], reverse=True),
+    StaticCxn(origin="Ruined Atoll", destination="Ruined Atoll Portal",
+              reqs=[["Prayer"]]),
+    StaticCxn(origin="Ruined Atoll Portal", destination="Ruined Atoll"),
+    StaticCxn(origin="Frog's Domain", destination="Frog's Domain Back",
+              reqs=[["Magic Orb"]]),
+    StaticCxn(origin="Library Exterior Tree", destination="Library Exterior Ladder",
+              reqs=[["Magic Orb"], ["Hero's Laurels"]]),
+    StaticCxn(origin="Library Exterior Ladder", destination="Library Exterior Tree",
+              reqs=[["Magic Orb", "Prayer"], ["Hero's Laurels", "Prayer"]]),
+    StaticCxn(origin="Library Hall", destination="Library Hero's Grave",
+              reqs=[["Prayer"]]),
+    StaticCxn(origin="Library Hero's Grave", destination="Library Hall"),
+    StaticCxn(origin="Library Lab Lower", destination="Library Lab",
+              reqs=[["Hero's Laurels"], ["Magic Orb"]]),
+    StaticCxn(origin="Library Lab", destination="Library Lab Lower",
+              reqs=[["Hero's Laurels"]]),
+    StaticCxn(origin="Library Lab", destination="Library Portal",
+              reqs=[["Prayer"]]),
+    StaticCxn(origin="Library Portal", destination="Library Lab"),
+    StaticCxn(origin="Fortress Exterior from East Forest", destination="Fortress Exterior from Overworld",
+              reqs=[["Hero's Laurels"], ["Magic Orb"]]),
+    StaticCxn(origin="Fortress Exterior from Overworld", destination="Fortress Exterior from East Forest",
+              reqs=[["Hero's Laurels"]]),
+    StaticCxn(origin="Fortress Exterior from Overworld", destination="Fortress Exterior near cave",
+              reqs=[["Hero's Laurels"], ["Prayer"]]),
+    StaticCxn(origin="Fortress Exterior near cave", destination="Fortress Exterior from Overworld",
+              reqs=[["Hero's Laurels"]]),
+    StaticCxn(origin="Fortress Exterior from Overworld", destination="Fortress Courtyard",
+              reqs=[["Hero's Laurels"]], reverse=True),
+    StaticCxn(origin="Fortress Courtyard Upper", destination="Fortress Courtyard"),
+    StaticCxn(origin="Fortress Courtyard Upper", destination="Fortress Exterior from Overworld"),
+    StaticCxn(origin="Beneath the Vault Front", destination="Beneath the Vault Back",
+              reqs=[["Lantern"]]),
+    StaticCxn(origin="Beneath the Vault Back", destination="Beneath the Vault Front"),
+    StaticCxn(origin="Fortress East Shortcut Upper", destination="Fortress East Shortcut Lower"),
+    StaticCxn(origin="Eastern Vault Fortress", destination="Eastern Vault Fortress Gold Door",
+              reqs=[["Prayer"]],
+              region_reqs=["Fortress Exterior from Overworld", "Fortress Courtyard Upper",
+                           "Beneath the Vault Back", "Eastern Vault Fortress"], reverse=True),
+    StaticCxn(origin="Fortress Grave Path", destination="Fortress Grave Path Dusty Entrance",
+              reqs=[["Hero's Laurels"]], reverse=True),
+    StaticCxn(origin="Fortress Grave Path", destination="Fortress Hero's Grave",
+              reqs=[["Prayer"]]),
+    StaticCxn(origin="Fortress Hero's Grave", destination="Fortress Grave Path"),
+    StaticCxn(origin="Fortress Arena", destination="Fortress Arena Portal",
+              reqs=[["Prayer"]],
+              region_reqs=["Fortress Exterior from Overworld", "Beneath the Vault Back", "Eastern Vault Fortress"]),
+    StaticCxn(origin="Fortress Arena Portal", destination="Fortress Arena"),
+    StaticCxn(origin="Lower Mountain", destination="Lower Mountain Stairs",
+              reqs=[["Holy Cross"]], reverse=True),
+    StaticCxn(origin="Quarry", destination="Lower Quarry",
+              reqs=[["Scavenger Mask", "Sword"], ["Scavenger Mask", "Magic Wand"]], reverse=True),
+    StaticCxn(origin="Quarry", destination="Quarry Portal",
+              reqs=[["Prayer", "Magic Orb"]],
+              region_reqs=["Quarry Connector"]),
+    StaticCxn(origin="Quarry Portal", destination="Quarry"),
+    StaticCxn(origin="Lower Quarry", destination="Lower Quarry Zig Door",
+              reqs=[["Magic Orb", "Prayer"]],
+              region_reqs=["Quarry Connector", "Quarry"], reverse=True),
+    StaticCxn(origin="Monastery Rope", destination="Quarry"),
+    StaticCxn(origin="Monastery Front", destination="Monastery Back"),
+    StaticCxn(origin="Monastery Back", destination="Monastery Front",
+              reqs=[["Hero's Laurels"]]),
+    StaticCxn(origin="Monastery Back", destination="Monastery Hero's Grave",
+              reqs=[["Prayer"]]),
+    StaticCxn(origin="Monastery Hero's Grave", destination="Monastery Back"),
+    StaticCxn(origin="Rooted Ziggurat Upper Front", destination="Rooted Ziggurat Upper Back",
+              reqs=[["Hero's Laurels"], ["Sword"]]),
+    StaticCxn(origin="Rooted Ziggurat Upper Back", destination="Rooted Ziggurat Upper Front",
+              reqs=[["Hero's Laurels"]]),
+    StaticCxn(origin="Rooted Ziggurat Middle Top", destination="Rooted Ziggurat Middle Bottom"),
+    StaticCxn(origin="Rooted Ziggurat Lower Front", destination="Rooted Ziggurat Lower Back",
+              reqs=[["Hero's Laurels"], ["Sword", "Prayer"]]),
+    StaticCxn(origin="Rooted Ziggurat Lower Back", destination="Rooted Ziggurat Lower Front",
+              reqs=[["Hero's Laurels"]]),
+    StaticCxn(origin="Rooted Ziggurat Lower Back", destination="Rooted Ziggurat Portal Room Entrance",
+              reqs=[["Prayer"]]),
+    StaticCxn(origin="Rooted Ziggurat Portal Room Entrance", destination="Rooted Ziggurat Lower Back"),
+    StaticCxn(origin="Rooted Ziggurat Portal", destination="Rooted Ziggurat Portal Room Exit",
+              reqs=[["Prayer"]],
+              region_reqs=["Rooted Ziggurat Lower Back"]),
+    StaticCxn(origin="Rooted Ziggurat Portal Room Exit", destination="Rooted Ziggurat Portal",
+              reqs=[["Prayer"]]),
+    StaticCxn(origin="Swamp", destination="Swamp to Cathedral Main Entrance",
+              reqs=[["Prayer"]],
+              region_reqs=["Overworld Laurels", "Swamp"], reverse=True),
+    StaticCxn(origin="Swamp", destination="Swamp to Cathedral Treasure Room",
+              reqs=[["Holy Cross"]]),
+    StaticCxn(origin="Swamp to Cathedral Treasure Room", destination="Swamp"),
+    StaticCxn(origin="Back of Swamp", destination="Back of Swamp Laurels Area",
+              reqs=[["Hero's Laurels"]], reverse=True),
+    StaticCxn(origin="Back of Swamp", destination="Swamp Hero's Grave",
+              reqs=[["Prayer"]]),
+    StaticCxn(origin="Swamp Hero's Grave", destination="Back of Swamp"),
+    StaticCxn(origin="Cathedral Gauntlet Checkpoint", destination="Cathedral Gauntlet"),
+    StaticCxn(origin="Cathedral Gauntlet", destination="Cathedral Gauntlet Exit",
+              reqs=[["Hero's Laurels"]], reverse=True),
+    StaticCxn(origin="Far Shore", destination="Far Shore to Spawn",
+              reqs=[["Hero's Laurels"]], reverse=True),
+    StaticCxn(origin="Far Shore", destination="Far Shore to East Forest",
+              reqs=[["Hero's Laurels"]], reverse=True),
+    StaticCxn(origin="Far Shore", destination="Far Shore to West Garden",
+              reqs=[["Prayer"]],
+              region_reqs=["West Garden"]),
+    StaticCxn(origin="Far Shore to West Garden", destination="Far Shore"),
+    StaticCxn(origin="Far Shore", destination="Far Shore to Quarry",
+              reqs=[["Prayer", "Magic Orb"]],
+              region_reqs=["Quarry", "Quarry Connector"]),
+    StaticCxn(origin="Far Shore to Quarry", destination="Far Shore"),
+    StaticCxn(origin="Far Shore", destination="Far Shore to Fortress",
+              reqs=[["Prayer"]],
+              region_reqs=["Fortress Exterior from Overworld", "Beneath the Vault Back", "Eastern Vault Fortress"]),
+    StaticCxn(origin="Far Shore to Fortress", destination="Far Shore"),
+    StaticCxn(origin="Far Shore", destination="Far Shore to Library",
+              reqs=[["Prayer"]],
+              region_reqs=["Library Lab"]),
+    StaticCxn(origin="Far Shore to Library", destination="Far Shore"),
+]
+
+
+# key is the region you have, value is the regions you get for having it
+# so that we aren't being excessively careful with the granular regions
+dependent_regions: Dict[Tuple[str, ...], List[str]] = {
+    ("Overworld", "Overworld Belltower", "Overworld Laurels", "Overworld Southeast Cross Door", "Overworld Temple Door",
+     "Overworld Fountain Cross Door", "Overworld Town Portal", "Overworld Spawn Portal"):
+         ["Overworld", "Overworld Belltower", "Overworld Laurels", "Overworld Ruined Hall Door",
+          "Overworld Southeast Cross Door", "Overworld Old House Door", "Overworld Temple Door",
+          "Overworld Fountain Cross Door", "Overworld Town Portal", "Overworld Spawn Portal"],
+    ("Old House Front",): ["Old House Front", "Old House Back"],
+    ("Furnace Fuse", "Furnace Ladder Area", "Furnace Walking Path"):
+        ["Furnace Fuse", "Furnace Ladder Area", "Furnace Walking Path"],
+    ("Sealed Temple", "Sealed Temple Rafters"): ["Sealed Temple", "Sealed Temple Rafters"],
+    ("Forest Belltower Upper",): ["Forest Belltower Upper", "Forest Belltower Main", "Forest Belltower Lower"],
+    ("Forest Belltower Main",): ["Forest Belltower Main", "Forest Belltower Lower"],
+    ("East Forest", "East Forest Dance Fox Spot", "East Forest Portal"):
+        ["East Forest", "East Forest Dance Fox Spot", "East Forest Portal"],
+    ("Forest Grave Path Main", "Forest Grave Path Upper"):
+        ["Forest Grave Path Main", "Forest Grave Path Upper",
+         "Forest Grave Path by Grave", "Forest Hero's Grave"],
+    ("Forest Grave Path by Grave", "Forest Hero's Grave"):
+        ["Forest Grave Path by Grave", "Forest Hero's Grave"],
+    ("Bottom of the Well Front", "Bottom of the Well Back"): ["Bottom of the Well Front", "Bottom of the Well Back"],
+    ("Dark Tomb Entry Point", "Dark Tomb Main", "Dark Tomb Dark Exit"):
+        ["Dark Tomb Entry Point", "Dark Tomb Main", "Dark Tomb Dark Exit"],
+    ("Dark Tomb Checkpoint", "Well Boss"): ["Dark Tomb Checkpoint", "Well Boss"],
+    ("West Garden", "West Garden Laurels Exit", "West Garden after Boss", "West Garden Hero's Grave"):
+        ["West Garden", "West Garden Laurels Exit", "West Garden after Boss", "West Garden Hero's Grave"],
+    ("Ruined Atoll", "Ruined Atoll Lower Entry Area", "Ruined Atoll Frog Mouth", "Ruined Atoll Portal"):
+        ["Ruined Atoll", "Ruined Atoll Lower Entry Area", "Ruined Atoll Frog Mouth", "Ruined Atoll Portal"],
+    ("Frog's Domain",): ["Frog's Domain", "Frog's Domain Back"],
+    ("Library Exterior Ladder", "Library Exterior Tree"):
+        ["Library Exterior Ladder", "Library Exterior Tree"],
+    ("Library Hall", "Library Hero's Grave"): ["Library Hall", "Library Hero's Grave"],
+    ("Library Lab", "Library Lab Lower", "Library Portal"): ["Library Lab", "Library Lab Lower", "Library Portal"],
+    ("Fortress Courtyard Upper",):
+        ["Fortress Courtyard Upper", "Fortress Exterior from East Forest", "Fortress Exterior from Overworld",
+         "Fortress Exterior near cave", "Fortress Courtyard"],
+    ("Fortress Exterior from East Forest", "Fortress Exterior from Overworld",
+     "Fortress Exterior near cave", "Fortress Courtyard"):
+        ["Fortress Exterior from East Forest", "Fortress Exterior from Overworld",
+         "Fortress Exterior near cave", "Fortress Courtyard"],
+    ("Beneath the Vault Front", "Beneath the Vault Back"): ["Beneath the Vault Front", "Beneath the Vault Back"],
+    ("Fortress East Shortcut Upper",): ["Fortress East Shortcut Upper", "Fortress East Shortcut Lower"],
+    ("Eastern Vault Fortress", "Eastern Vault Fortress Gold Door"):
+        ["Eastern Vault Fortress", "Eastern Vault Fortress Gold Door"],
+    ("Fortress Grave Path", "Fortress Grave Path Dusty Entrance", "Fortress Hero's Grave"):
+        ["Fortress Grave Path", "Fortress Grave Path Dusty Entrance", "Fortress Hero's Grave"],
+    ("Fortress Arena", "Fortress Arena Portal"): ["Fortress Arena", "Fortress Arena Portal"],
+    ("Lower Mountain", "Lower Mountain Stairs"): ["Lower Mountain", "Lower Mountain Stairs"],
+    ("Monastery Front", "Monastery Back", "Monastery Hero's Grave"):
+        ["Monastery Front", "Monastery Back", "Monastery Hero's Grave"],
+    ("Quarry", "Quarry Portal", "Lower Quarry"): ["Quarry", "Quarry Portal", "Lower Quarry", "Lower Quarry Zig Door"],
+    ("Monastery Rope",): ["Monastery Rope", "Quarry", "Quarry Portal", "Lower Quarry", "Lower Quarry Zig Door"],
+    ("Rooted Ziggurat Upper Front", "Rooted Ziggurat Upper Back"):
+        ["Rooted Ziggurat Upper Front", "Rooted Ziggurat Upper Back"],
+    ("Rooted Ziggurat Middle Top",): ["Rooted Ziggurat Middle Top", "Rooted Ziggurat Middle Bottom"],
+    ("Rooted Ziggurat Lower Front", "Rooted Ziggurat Lower Back", "Rooted Ziggurat Portal Room Entrance"):
+        ["Rooted Ziggurat Lower Front", "Rooted Ziggurat Lower Back", "Rooted Ziggurat Portal Room Entrance"],
+    ("Rooted Ziggurat Portal", "Rooted Ziggurat Portal Room Exit"):
+        ["Rooted Ziggurat Portal", "Rooted Ziggurat Portal Room Exit"],
+    ("Swamp", "Swamp to Cathedral Treasure Room", "Swamp to Cathedral Main Entrance"):
+        ["Swamp", "Swamp to Cathedral Treasure Room", "Swamp to Cathedral Main Entrance"],
+    ("Back of Swamp", "Back of Swamp Laurels Area", "Swamp Hero's Grave"):
+        ["Back of Swamp", "Back of Swamp Laurels Area", "Swamp Hero's Grave"],
+    ("Cathedral Gauntlet Checkpoint",):
+        ["Cathedral Gauntlet Checkpoint", "Cathedral Gauntlet Exit", "Cathedral Gauntlet"],
+    ("Far Shore", "Far Shore to Spawn", "Far Shore to East Forest", "Far Shore to Quarry",
+     "Far Shore to Fortress", "Far Shore to Library", "Far Shore to West Garden"):
+        ["Far Shore", "Far Shore to Spawn", "Far Shore to East Forest", "Far Shore to Quarry",
+         "Far Shore to Fortress", "Far Shore to Library", "Far Shore to West Garden"]
+}
